+++ apps/web/src/components/Layout/MainLayout.tsx
@@ -7,8 +7,6 @@ import useLayoutTheme from "~/hooks/useLayoutTheme";
 
 import Layout from "./Layout";
 
-export { ErrorBoundary } from "./ErrorBoundary/ErrorBoundary";
-
 export const Component = () => {
   const intl = useIntl();
   useLayoutTheme("default");
