+++ apps/web/src/components/Layout/RouteErrorBoundary/RouteErrorBoundary.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryFn } from "@storybook/react";
 
 import { CHROMATIC_VIEWPORTS } from "@gc-digital-talent/storybook-helpers";
 
-import { ErrorBoundary } from "./ErrorBoundary";
+import ErrorBoundary from "./RouteErrorBoundary";
 
 export default {
   component: ErrorBoundary,
