+++ apps/web/src/components/Layout/RouteErrorBoundary/RouteErrorBoundary.tsx
@@ -12,7 +12,7 @@ import useErrorMessages from "~/hooks/useErrorMessages";
 import darkPug from "~/assets/img/404_pug_dark.webp";
 import lightPug from "~/assets/img/404_pug_light.webp";
 
-export const ErrorBoundary = () => {
+export const RouteErrorBoundary = () => {
   const intl = useIntl();
   const paths = useRoutes();
   const { mode } = useTheme();
@@ -119,4 +119,4 @@ export const ErrorBoundary = () => {
   );
 };
 
-export default ErrorBoundary;
+export default RouteErrorBoundary;
